backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.1
pydantic-settings==2.1.0
python-ldap==3.4.3
PyJWT==2.8.0
fpdf2==2.7.5
python-dotenv==1.0.0
SQLAlchemy==2.0.23
starlette-rate-limit==0.5.0
fastapi-mail==1.4.1
python-magic==0.4.27
tenacity==8.2.3
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1